move whiteboard code to seperate file
